Logger: Setter and getter test
It looks like the log level is currently not properly set, thus causing
the output messages to be output regardless the actual requested logging
level. Such a test may be trivial for setter and getter classes, but can
be slightly trickier in the case of singletons.

Implement a test that tries to set the logger level and reads it back,
to ensure the functionality is correct.

Bug: 194172256
Test: atest dittobench_test
Signed-off-by: Alessio Balsini <balsini@google.com>
Change-Id: I557d977aa41b9a4b3d0d28d2bd00e87b451a66f4