Implement shared variables
Implement shared_variables as a static vector, which is accessible to
all the instructions. Instructions are able to read and write a specific
variable by providing an integer key. Each integer key is matched with a
string name, that is provided in the .ditto as part of an instruction.
Having a map that matches string variable names, that are inputted by
the user, with integer keys, that can then be used to access the shared
variables during the benchmark, allows for better performance, since the
map is only maintained during the setup, while the actual values are
stored in the vector and are accessed with integer keys during the
benchmark.

Bug: 194484941
Test: dittobench_test
Signed-off-by: Robertas Norkus <robertasn@google.com>
Signed-off-by: Alessio Balsini <balsini@google.com>
Change-Id: Ieece31655003ab0c93d6ec994aeae84b01a80401