Implement WriteFile instruction
Implement WriteFile instruction, which takes file descriptor,
size, block size, type (sequential, random) and seed (for random).
In SetUp() instruction calculates all the units (size and offset)
to which a buffer must be written and then in Run() they are written.
This choice to do the computations (which involve random number
generation) in SetUp() before Run() allows to exclude the duration
of executing these computations from the duration of the whole
execution of instruction, although it still contributes to the
execution duration of the parent instruction set. It is not possible
to move these computations outside the benchmark, since the file
size gets only known after the start of the benchmark. The solution
on how to fix this issue will be considered in the future and
refactoring will take place if needed.

Bug: 194366184
Test: manual
Signed-off-by: Robertas Norkus <robertasn@google.com>
Change-Id: I07a6c9ce6e2f4fc90b7b3333a30b5d482ff30394