Refactor build configurations with *.cpp
Adding every source to both the Android.bp and CMakeLists.txt was
becoming time consuming and these two files would have required to be
always consistent.
Fix by building all the files in src/ as a library called
"libdittosuite".  Also move any source files containing a main() to the
root of the project (dittobench.cpp) as they are not part of the
library.

Bug: 194679286
Test: m, cmake
Signed-off-by: Alessio Balsini <balsini@google.com>
Change-Id: I39186f1dee2ed2e4bffadbda011b451a54893342