Add option for ReadWriteFile to reseed random
Add option for ReadWriteFile to reseed random.
If ONCE is set, random number generator is seeded only
during the construction of instruction. If EACH_ROUND_OF_CYCLES
is set, then random number generator is reseeded in SetUp() method.
If EACH_CYCLE is set, then the generator is reseeded in SetUpSingle()
method. This optional allows for greater control of random offset
generation in Read and Write file instruction.

Bug: 195388141
Test: manual
Signed-off-by: Robertas Norkus <robertasn@google.com>
Change-Id: I38e6f3827757555a45f6bb9052a742439cb733f0