Implement DittoTrace
DittoTrace takes in strace output as input and outputs .ditto file.
Also, takes in absolute path as additional parameter, that
determines which files should only be considered.

Processing:
It reads all the lines from the strace output, splits them by pid
for easier parsing, goes over each line in each pid and concatenates
"unfinished" and "resumed" lines, then processes each line into
a syscall name, arguments and return value.

Execution:
Goes through processed lines one by one, creates an instruction set
with an open file instruction when openat() with unseen fd is spotted.
Then, if pread64() or pwrite64() is spotted for the fd which was
opened before, read file or write file instruction is created
accordingly. Finally, when close() is spotted for the fd which was
opened before, the instruction set for this fd is closed and added
to the main instruction set.

Bug: 200543075
Test: manual
Signed-off-by: Robertas Norkus <robertasn@google.com>
Change-Id: Ibd4c42abf923cbe5bb1239a694a07d80cb78e759