Added statistics table output
Method PrintStatisticsTable prints a table containing
information regarding the statistics (every instruction
path/name, with the corresponding min, max, mean).

The first column of the table represents the instruction
path. It displays up to 2 'parent' instructions in the
instruction tree. For example,
instructionA/instructionB/instructionC/instructionD
will be printed as
..instructionC/instructionD
and the same for longer instruction paths.

Bug: 195113426
Test: manual
Signed-off-by: Lucia Lup <lucialup@google.com>
Change-Id: I5f6391d44e24ed37854e2899930c54b0aa8b4a76