Histogram output
Added histogram output. It prints a horizontal normalized
histogram according to the statistics of the instruction
and its subinstructions (currently available statistics
are min, max, mean).

The histogram output is normalized on both axis (on the
Oy axis, it uses the constant kBinSize to define the bin
size for the time samples, and on the Ox axis, it uses
the constant kMaxHistogramWidth to define the maximum
width of the histogram).

The time sample axis of the histogram (Ox) is represented
in nanoseconds.

Bug: 195113426
Test: manual
Signed-off-by: Lucia Lup <lucialup@google.com>
Change-Id: I5c30cf8e32cc63da46f502033cb362a8ea4c441d