Change time samples type in Result class
While collecting the time samples for an instruction,
the results were previously passed to the Result class
as a vector of timespecs. The latter is changed into
a vector of long integers, so the entire Result class
handles long integers instead of timespecs (which is
helpful for making it more generic in the future).

The output is not affected by this change, as all
timespecs were printed as long integers already.

Bug: 195113426
Test: manual
Signed-off-by: Lucia Lup <lucialup@google.com>
Change-Id: I84bedb187dc92eecf14dfcb22c8b22f0ebd404bb