Introduce AddMeasurement for Result class
The scope of this change is to accommodate different
measurements in the Result class.

Result class contains a map between the name of the
measured parameter and the vector of samples for that
measuring (for example, a map between "duration" and
a vector of time samples). AddMeasurement method adds
a new entry in the samples map containing the key (the
string name) and value (the samples long int vector)
given as parameters.

The printing functions are refactored in order to
support the new change.

Bug: 195113227
Test: manual
Signed-off-by: Lucia Lup <lucialup@google.com>
Change-Id: I08aba50578b00c133ccb9d9cd0fa57abb0de098e