Collect bandwidth for read/write instructions
BandwidthSampler is a subclass of the generic Sampler class,
measuring bandwith depending on the file size and duration
given as parameters.

For read/write instructions, the bandwidth is measured at
TearDownSingle (where the size and the duration of the
current instruction are available). While collecting the
result for this instructions, the bandwidth measurement
is also added to the result class.

Bug: 196960127
Test: manual
Signed-off-by: Lucia Lup <lucialup@google.com>
Change-Id: I8d0e9de812cd2977f8e5efd41728e76071cc4190