Fix bug where block size was not set as file size
When -1 is provided as the block size for read and write operations,
it is expected for the block size to take the value of file size.
Buffer used to be initialized with the size of the block size in the
constructor, where the block size was still set to -1, so the binary
crashed.

Now the buffer is initialized during the setup when the file
size becomes known. Also, bad_alloc is now handled.

In addition, tests that can catch this bug were added in this change.

Fix: 6fdc60dcadc9 ("Add option to use file size as block size")
Bug: 196542737
Test: dittobench_test
Signed-off-by: Robertas Norkus <robertasn@google.com>
Change-Id: I5ba8e0f85e907f1297f8befe59c16b24b4e3c04b