Fix bug where size did not match file size when repeated
When -1 was provided as size or block size for Read/Write,
size and block size got set to the current file size. If the
instruction was repeated on a different file, the size and block size
stayed the same and were not set to a new file size.

Now size and block size are only set locally in the SetupSingle(),
so that -1 is not overwritten.

In addition, tests that can catch this bug were added in this change.

Bug: 193388430
Test: dittobench_test
Signed-off-by: Robertas Norkus <robertasn@google.com>
Change-Id: I627ca9a36286bf5fbf51b496378cb6c366eaedd3