Change purpose of Result Print() method
Print() method used to output for every instruction
its measurements statistics. As this information
is also presented in the statistics table and in
the csv output format, this functionality is not
needed.

However, in order to keep the code clean, the
Print() method now handles the whole printing of
the results (it calls the methods printing the
histograms, tables and csv). This way, the methods
that print the histograms, tables and csv can be
made private.

Bug: 195113426
Test: manual
Signed-off-by: Lucia Lup <lucialup@google.com>
Change-Id: I268a0413d437eff2dbe60952490e50e3280904bc