Use double samples in AddMeasurement
AddMeasurement method from Result class adds a new
entry in the samples map (the key being the name of
the measurement while the value is the measurement
samples vector).

Samples data type is changed from long int to
double in order to provide a better accuracy.
Keeping samples as doubles improves the precision
when computing the measurements statistics.

Bug: 195113227
Test: manual
Signed-off-by: Lucia Lup <lucialup@google.com>
Change-Id: I87e8e05396ed258356151dcb02e8d238297c3acc