Rename AccessType enumerator to Order
AccessType is intended as an enumerator to specify the order at which
certain data should be accessed. This type should be not limited to
specific data types, for example files or elements in a list.
For the specific case of access to files, the AccessType is misleading
as might be confused to file open modes.
Rename AccessType to Order would solve this name ambiguity.

Test: build
Signed-off-by: Alessio Balsini <balsini@google.com>
Change-Id: Ifa10ffd71ef1a09e26dea168f5ee3ca4997d2b3a