Fix proto header export for new Soong rules
Some recent changes in Soong are causing the generated headers from
Protocol Buffers to be not found.

The Android.bp documentation is not clear for this specific flag, but as
far as I can see from the build log, when this flag is set to `false`,
the generated files are put into a generic `out/.../proto` directory
relative to the project, but when set to `true`, to the generated
headers directory becomes `out/.../proto/<project path>`.
The result is tat when this flag is true, the generated headers path
depends on where the source of this project is within the Android code
base.
For example, I moved this project from `$ANDROID_BUILD_TOP/test/` to
`$ANDROID_BUILD_TOP/external/` and the build was failing.

These new adjustments to the Android.bp files fix the issue.

Bug: 243896359
Test: mma .
Signed-off-by: Alessio Balsini <balsini@google.com>
Change-Id: I98ead807a2365bc1192e23ef05bc06c4126cb2bf