Add parameter to TearDownSingle
Some instructions may benefit from knowing what is the last iteration.
For example, in the case of producer-consumer, the consumer needs a
signal to be shut down. This can be done in multiple ways, but the
design chosen here is the client to send a known command to terminate
the server when at the last iteration.

Test: ctest
Bug: 303866491
Change-Id: Ib39239aa533ef05dea7e5546b6febf2450d1b892
Signed-off-by: Alessio Balsini <balsini@google.com>