Microseconds conversion utility
Nanoseconds can be good to measure time to limit
possible propagation errors, but when it comes to describe durations
from the users, nanoseconds require too many figures.
As right now a syscall takes something in the order of the microseconds,
this can be a handy measurement unit to be added.

Bug: 304826216
Test: ctest
Change-Id: Ia9142d109de8b5fea5f32f00055d07d9a8150d6b
Signed-off-by: Alessio Balsini <balsini@google.com>