Fix naming of threads
The name of processes was already correctly set, but the name of threads
was still missing.
This patch fixes the issues, so that now also threads are set the
correct name.
In order to do so, apply also a small refactoring of the way parameters
of threads/processes are set, by using a separate struct.

Bug: 305023314
Test: manual with Perfetto
Change-Id: Ie955ec2795740935b2087efb0805b1dbc7f82862
Signed-off-by: Alessio Balsini <balsini@google.com>