Introduce Bazel build
Create all the build configuration to be able to build dittosuite with
Bazel.
The tricky part is the Protocol Buffer dependency, which is solved in
the MODULE.bazel file. This works fine if the project is built
standalone, but when imported as part of another Bazel project, the
WORKSPACE dependency is not met, and the build fails.

Test: bazel build :all
Bug: 322744630
Change-Id: Icaeffb26b5251d1ff33755eb79af528cfb2c8d58
Signed-off-by: Alessio Balsini <balsini@google.com>