Move gettid and sched_setattr to Syscall interface
Moving function calls that are strictly depending on the platform into
an external redirection that complies to the SyscallInterface interface
is a design pattern that should have been always followed.
Not complying to this design pattern can cause issues, for example when
building the tool with Kleaf, whose build environment does not support
some of the latest syscalls.
Fix by moving these function calls into the Syscall interface and
implementing them in a more baremetal fashion.

Bug: 322744630
Test: build, run on Android
Change-Id: I2e9806373868fadd54f9d500346aaab424e34539
Signed-off-by: Alessio Balsini <balsini@google.com>