Update protobuf header include path
When building the project with Kleaf, the root path is different from
the Dittosuite root, causing the protobuffer-generated headers to not be
found in the expected folder.
After several attemps, the simplest solution that unifies Android.bp,
CMake and Bazel is to explicitly include the PB directory "schema". This
works for both CMake and Bazel, and is also what was is done by Android
by default.
The `#ifdef __ANDROID__` condition to choose if including
`<benchmark.pb.h>` or `"schema/benchmark.pb.h"` is not needed anymore,
as `<benchmark.pb.h>` now works in both cases.

Bug: 322744630
Test: bazel build --enable_bzlmod --config=internet \
  //test/dittosuite:dittobench # From common-android-mainline
Change-Id: Ibb90679689178169add35c866c4935f30686ccc0
Signed-off-by: Alessio Balsini <balsini@google.com>