Use name instead of instruction id for spawned tasks
When printing the benchmarking results, instead of reporting the
process/thraed name, Dittobench was only printing the numerical
identifier of the process/thread. This was enough to distinguish the
instructions run from one process/thread to another, but not enough to
understand what process/thread called them.
Additionally, this numerical value was only referring to the id of the
"spawned" processes, so if multiple threads were generated manually, and
each of them was using the default "spawn" value of 1, their reported
instruction identifier would have been "0".
Fix by using the thread name if available, otherwise the instruction
number.

Test: manual check
Change-Id: If3cfbf705dd242f78a51cec995381d964872217f
Signed-off-by: Alessio Balsini <balsini@google.com>