Embed all the example workloads within the DittoBench binary
To make DittoBench completely standalone, modify the build scripts to
generate a source file that includes all the workloads defined in the
example folder. These workloads can then be used without needing the
associated .ditto file. A workload can still be provided as a .ditto
file, same as before.
The embedded workload can be chosen with the new command line argument
"-w" ("--workload").
The list of available workloads is presented to the user with the binary
help, passing the additional optional parameter "-h workloads".

Bug: 343217933
Test: build and run dittobench, both with file and embedded benchmarks
Change-Id: Ic109dce225015eb59aa1939f6075b46b973760e6
Signed-off-by: Alessio Balsini <balsini@google.com>