Test all the ditto files in example/ recursively
Extend the testing of .ditto files (parsing and creation of all the
object instances in the factory method) not only to the contents of
example/, but also to recurse the whole example/ directory.
For targets that are not Android, the Android-specific ditto files are
skipped, meaning that the test report will still show that those tests
could not execute, but without failing.

Bug: 343704062
Test: ctest --test-dir test
Change-Id: I41d97a635bf7b61c1d25aa46c384d41b1f92deb8
Signed-off-by: Alessio Balsini <balsini@google.com>